+++ py/desimodel/test/test_footprint.py
@@ -17,7 +17,42 @@ class TestFootprint(unittest.TestCase):
     
     def setUp(self):
         io.reset_cache()
-            
+
+    def test_pass2program(self):
+        '''Test footprint.pass2program(tilepass)
+        '''
+        self.assertEqual(footprint.pass2program(0), 'DARK')
+        self.assertEqual(footprint.pass2program(1), 'DARK')
+        self.assertEqual(footprint.pass2program(2), 'DARK')
+        self.assertEqual(footprint.pass2program(3), 'DARK')
+        self.assertEqual(footprint.pass2program(4), 'GRAY')
+        self.assertEqual(footprint.pass2program(5), 'BRIGHT')
+        self.assertEqual(footprint.pass2program(6), 'BRIGHT')
+        self.assertEqual(footprint.pass2program(7), 'BRIGHT')
+
+        passes = [0,1,2,3,4,5,6,7]
+        programs = ['DARK', 'DARK', 'DARK', 'DARK', 'GRAY', 'BRIGHT', 'BRIGHT', 'BRIGHT']
+        tmp = footprint.pass2program(passes)
+        self.assertEqual(tmp, programs)
+
+        tmp = footprint.pass2program(np.array(passes))
+        self.assertEqual(tmp, programs)
+
+        tmp = footprint.pass2program([0,0,1])
+        self.assertEqual(tmp, ['DARK', 'DARK', 'DARK'])
+
+        with self.assertRaises(KeyError):
+            footprint.pass2program(999)
+
+    def test_program2pass(self):
+        '''Test footprint.program2pass()
+        '''
+        self.assertEqual(footprint.program2pass('DARK'), [0,1,2,3])
+        self.assertEqual(footprint.program2pass('GRAY'), [4,])
+        self.assertEqual(footprint.program2pass('BRIGHT'), [5,6,7])
+        with self.assertRaises(ValueError):
+            footprint.program2pass('BLAT')
+
     def test_get_tile_radec(self):
         """Test grabbing tile information by tileID.
         """